Add function to create stories in utils
