+++ stories/ball.stories.tsx
@@ -1,6 +1,6 @@
 import React, { cloneElement } from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, boolean, color } from '@storybook/addon-knobs';
+import { withKnobs } from '@storybook/addon-knobs';
 
 import BallBeat from '../src/BallBeat';
 import BallClipRotate from '../src/BallClipRotate';
@@ -24,21 +24,7 @@ import BallTrianglePath from '../src/BallTrianglePath';
 import BallZigZag from '../src/BallZigZag';
 import BallZigZagDeflect from '../src/BallZigZagDeflect';
 
-const defaultColor = '#000000';
-const defaultState = true;
-
-const style = {
-    display: 'flex',
-    flex: '0 1 auto',
-    flexDirection: 'column' as 'column',
-    flexGrow: 1,
-    flexShrink: 0,
-    flexBasis: '25%',
-    maxWidth: '25%',
-    height: '150px',
-    alignItems: 'center',
-    justifyContent: 'center',
-};
+import createStories from '../utils/createStories';
 
 const animationsStories = [
     {
@@ -131,7 +117,4 @@ const stories = storiesOf('Ball Loaders', module);
 
 stories.addDecorator(withKnobs);
 
-animationsStories.forEach((animation) => {
-    const { name, component } = animation;
-    stories.add(name, () => (<div style={style}>{ cloneElement(component, { color: color('Color', defaultColor), loading: boolean('Loading', defaultState) }) }</div>));
-});
+createStories(animationsStories, stories);
