+++ stories/line.stories.tsx
@@ -1,6 +1,6 @@
 import React, { cloneElement } from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, color, boolean } from '@storybook/addon-knobs';
+import { withKnobs } from '@storybook/addon-knobs';
 
 import LineScale from '../src/LineScale';
 import LineScaleParty from '../src/LineScaleParty';
@@ -9,21 +9,7 @@ import LineScalePulseOutRapid from '../src/LineScalePulseOutRapid';
 import LineScaleRandom from '../src/LineScaleRandom';
 import LineSpinFadeLoader from '../src/LineSpinFadeLoader';
 
-const defaultColor = '#000000';
-const defaultState = true;
-
-const style = {
-    display: 'flex',
-    flex: '0 1 auto',
-    flexDirection: 'column' as 'column',
-    flexGrow: 1,
-    flexShrink: 0,
-    flexBasis: '25%',
-    maxWidth: '25%',
-    height: '150px',
-    alignItems: 'center',
-    justifyContent: 'center',
-};
+import createStories from '../utils/createStories';
 
 const animationsForStories = [
     {
@@ -56,7 +42,4 @@ const stories = storiesOf('Line Loaders', module);
 
 stories.addDecorator(withKnobs);
 
-animationsForStories.forEach((animation) => {
-    const { name, component } = animation;
-    stories.add(name, () => (<div style={style}>{ cloneElement(component, { color: color('Color', defaultColor), loading: boolean('Loading', defaultState) })}</div>));
-});
+createStories(animationsForStories, stories);
